requirements.txt
Django==2.1.4
djangorestframework==3.7.7
Pillow==5.3.0
pytz==2018.7
